feat(ray): unify cli commands
