+++ instill/helpers/init-templates/.dockerignore
