+++ instill/helpers/init-templates/Dockerfile
