+++ instill/helpers/init-templates/instill.yaml
@@ -1,0 +1,20 @@
+build:
+  # set to true if your model requires GPU
+  gpu: true
+
+  # python version, currently only support 3.11
+  python_version: "3.11"
+
+  # cuda version if `gpu` is set to true
+  # support 11.5, 11.6 ,11.7 ,11.7 ,12.1
+  cuda_version: "12.1"
+
+  # a list of python packages in the format of {package-name}=={version}
+  # python_packages:
+  #   - torch==2.3.1
+  #   - transformers==4.41.2
+
+  # a list of system packages from apt package manager
+  # system_packages:
+  #   - htop
+  #   - libglib2.0-0
